LayoutTests/editing/pasteboard/paste-without-nesting-expected.txt

This test checks that copy a selection and paste over the same selection does not create nested divs.

PASS confimedMarkup is 'Hello'
PASS confimedMarkup is 'Hello<br>world'
PASS confimedMarkup is '<div>Hello</div><div>world</div>'
PASS confimedMarkup is '<div>Hello</div><div>world</div>'
PASS confimedMarkup is '<div><b><i>Hello</i></b></div><div><b><i>world</i></b></div>'
PASS confimedMarkup is '<div style="text-align: center;"><div><font color="#ff0000"><b><i>hello</i></b></font></div><div><font color="#ff0000"><b><i>world</i></b></font></div></div>'
PASS successfullyParsed is true

TEST COMPLETE
